refactor: skillsData & sidebarData moved to utils folder
